twitter_clone/requirements.txt
argcomplete==1.9.2
astroid==1.5.3
Babel==2.5.0
base58==0.2.4
boto3==1.4.7
botocore==1.7.14
certifi==2017.7.27.1
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
csscompressor==0.9.5
Django==2.2.7
django-appconf==1.0.1
django-autocomplete-light==3.3.2
django-compressor==2.2
django-debug-toolbar==1.10.1
django-form-utils==1.0.3
django-haystack==2.8.1
django-image-cropping==1.0.4
django-localflavor==1.5.2
django-markdownx==2.0.21
django-nested-admin==3.1.2
django-phonenumber-field==1.3.0
django-s3direct==1.0.3
django-smuggler==0.8.0
django-storages==1.6.5
django-widget-tweaks==1.4.1
docutils==0.14
durationpy==0.5
easy-thumbnails==2.4.1
future==0.16.0
futures==3.1.1
hjson==3.0.0
idna==2.6
importlib-metadata==0.6
isort==4.2.15
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.16.1
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
natsort==5.1.0
numpy==1.14.1
olefile==0.44
opencv-python==3.4.0.12
path.py==11.5.0
phonenumberslite==8.7.1
Pillow==4.2.1
placebo==0.8.1
psycopg2==2.7.3.2
pylint==1.7.2
python-dateutil==2.6.1
python-dotenv==0.8.2
python-monkey-business==1.0.0
python-slugify==1.2.4
pytz==2017.2
PyYAML==3.12
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
s3transfer==0.1.11
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
toml==0.9.2
tqdm==4.15.0
troposphere==2.0.2
typed-ast==1.1.0
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12
Whoosh==2.7.4
wrapt==1.10.11
wsgi-request-logger==0.4.6
zappa==0.44.3
